Update models.py
renamed the Post and PostCategory models into Thread and ThreadCategory. The author is now based on the user management app, and a header_image field is added to enable uploading of images. Added as well the Comment model with fields author, thread, entry, and the creation and last update dates and times.